requirements.txt
fastapi==0.79.0
uvicorn==0.18.2
requests==2.28.1
python-dotenv==0.20.0
pydantic==1.9.1
python-dateutil==2.8.2
grpcio==1.47.0
grpcio-tools==1.47.0
protobuf3==0.2.1